cv_project/requirements.txt
Django==5.1
djangorestframework==3.14.0python-dotenv
